Move localstoreage file
